Moved error control to UIManager
